% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxonomyManage.R
\name{addTerm}
\alias{addTerm}
\title{#################################################################
This file has funtions for managing taxonomy in a tree structure
#################################################################
#################################################################
function to add a word to the taxonomy}
\usage{
addTerm(taxtree, child, path = NULL)
}
\arguments{
\item{taxtree}{- the taxonomy tree to which word needs to be added}

\item{child}{- word to be added to taxonomy tree}

\item{parent}{- parent to which term will be added as child,
if parent is NULL then word is added to Root of tree}
}
\value{
updated taxonomy tree
}
\description{
#################################################################
This file has funtions for managing taxonomy in a tree structure
#################################################################
#################################################################
function to add a word to the taxonomy
}

